# openapi.yaml
...
openapi: 3.0.3
info:
  title: rucQ API
  version: 0.1.0
  description: rucQ API
servers:
  - url: http://localhost:8080
    description: 開発環境
paths:
...
  /api/rooms:
    get:
      summary: 部屋の一覧を取得
      tags:
        - Rooms
      operationId: getRooms
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      summary: 部屋を作成
      tags:
        - Rooms
      operationId: postRoom
      parameters:
        - $ref: "#/components/parameters/X-Forwarded-User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostRoomRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/rooms/{room_id}:
    put:
      summary: 部屋を更新
      tags:
        - Rooms
      operationId: putRoom
      parameters:
        - $ref: "#/components/parameters/RoomId"
        - $ref: "#/components/parameters/X-Forwarded-User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostRoomRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/staffs:
    get:
      summary: 合宿係の一覧を取得
      tags:
        - Staffs
      operationId: getStaffs
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      summary: 合宿係を追加
      tags:
        - Staffs
      operationId: postStaff
      parameters:
        - $ref: "#/components/parameters/X-Forwarded-User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostStaffRequest"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: 合宿係を削除
      tags:
        - Staffs
      operationId: deleteStaff
      parameters:
        - $ref: "#/components/parameters/StaffId"
        - $ref: "#/components/parameters/X-Forwarded-User"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  parameters:
    X-Forwarded-User:
      name: X-Forwarded-User
      in: header
      description: ログインしているユーザーのtraQ ID（NeoShowcaseが自動で付与）
      schema:
        type: string
...
    RoomId:
      name: room_id
      in: path
      description: 部屋ID
      required: true
      schema:
        type: integer
    StaffId:
      name: staff_id
      in: query
      description: 合宿係のtraQ ID
      required: true
      schema:
        type: string
  responses:
    NoContent:
      description: No Content
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Internal server error
  schemas:
...
    User:
      type: object
      properties:
        traq_id:
          type: string
        is_staff:
          type: boolean
      required:
        - traq_id
        - is_staff
...
    Room:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        camp_id:
          type: integer
        members:
          type: array
          items:
            $ref: "#/components/schemas/User"
      required:
        - id
        - name
        - camp_id
        - members
    PostRoomRequest:
      type: object
      properties:
        name:
          type: string
        camp_id:
          type: integer
        members:
          type: array
          items:
            type: string
      required:
        - name
        - camp_id
        - members
    PostStaffRequest:
      type: object
      properties:
        traq_id:
          type: string
      required:
        - traq_id
...